mvts 2

+++ src/hero.hpp
@@ -17,7 +17,7 @@ class Hero : public Character
     Hero(const irr::io::path&  filename, const irr::io::path&  textname, irr::core::vector3di position, irr::core::vector3df direction, float life, float baseDamage);
     Hero(const irr::io::path&  filename, const irr::io::path&  textname);
 
-    void compute_movements(int x_dest, int y_dest);
+    void compute_movements(float x_dest, float y_dest);
     void move();
 
     void compute_mvt_to_grid(irr::core::vector3df pos, Node_Graph closest_start);
